+++ src/index.ts
@@ -175,9 +175,18 @@ Hooks.on('ready', async () => {
             requiresReload: true,
             type: Boolean,
         } as any);
+        gameInstance.settings.register('pf2e-kingmaker-tools', 'kingdomAlwaysAddHalfLevel', {
+            name: 'Always add half Level to Skill',
+            hint: 'If enabled, always adds half of the kingdom\'s level to a skill, even if it is untrained',
+            scope: 'world',
+            config: true,
+            default: false,
+            requiresReload: true,
+            type: Boolean,
+        } as any);
         gameInstance.settings.register('pf2e-kingmaker-tools', 'kingdomAlwaysAddLevel', {
-            name: 'Always add Level to Skill',
-            hint: 'If enabled, always adds the kingdom\'s level to a skill, even if it is untrained',
+            name: 'Always add Level to Skill (Overrides Half Level)',
+            hint: 'If enabled, always adds the kingdom\'s level to a skill, even if it is untrained. Overrides ',
             scope: 'world',
             config: true,
             default: false,